requirements.txt
python-telegram-bot==13.7
python-decouple==3.4
pytesseract==0.3.8
Pillow==8.3.2
